+++ src/ParkMiller.h
@@ -8,22 +8,22 @@ using namespace std;
 class ParkMiller
 {
 private:
-    long seed;
+    unsigned long long int seed;
     static int A;
     static int M;
     static int Q;
     static int R;
 public:
     ParkMiller();
-    ParkMiller(long seed);
+    ParkMiller(unsigned long long int seed);
     ParkMiller(const ParkMiller & parkMiller);
     ParkMiller &operator= (const ParkMiller & parkMiller);
 
-    long get_seed() const;
-    void set_seed(long seed);
+    unsigned long long int get_seed() const;
+    void set_seed(unsigned long long int seed);
     void reset_seed();
 
-    long genererProchain();
+    unsigned long long int genererProchain();
 };
 
 #endif
